sysom_api/requirements.txt
aioredis==1.3.1
APScheduler==3.8.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
bcrypt==3.2.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-redis==5.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
paramiko==2.8.0
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.3
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.19.1
webencodings==0.5.1
zope.interface==5.4.0
